Add utility removePromiseCalls
