+++ src/transforms/v2-to-v3/transformer.ts
@@ -1,4 +1,4 @@
-import { API, FileInfo, Identifier } from "jscodeshift";
+import { API, FileInfo } from "jscodeshift";
 
 import {
   addV3ClientImport,
@@ -6,6 +6,7 @@ import {
   getV2ClientNames,
   getV2DefaultImportName,
   removeDefaultImportIfNotUsed,
+  removePromiseCalls,
   replaceClientCreation,
 } from "./utils";
 
@@ -24,49 +25,8 @@ export default function transformer(file: FileInfo, api: API) {
   for (const [v2ClientName, v3ClientMetadata] of Object.entries(clientMetadata).reverse()) {
     const { v3ClientName, v3ClientPackageName } = v3ClientMetadata;
     addV3ClientImport(j, source, { v3ClientName, v3ClientPackageName });
-
-    // ToDo: Move to utils file
-    source
-      .find(j.VariableDeclarator, {
-        id: { type: "Identifier" },
-        init: {
-          type: "NewExpression",
-          callee: {
-            object: { type: "Identifier", name: v2DefaultImportName },
-            property: { type: "Identifier", name: v2ClientName },
-          },
-        },
-      })
-      .forEach((nodePath) => {
-        const name = (nodePath.value.id as Identifier).name;
-        source
-          .find(j.CallExpression, {
-            callee: {
-              object: {
-                type: "CallExpression",
-                callee: {
-                  type: "MemberExpression",
-                  object: {
-                    type: "Identifier",
-                    name,
-                  },
-                },
-              },
-              property: { type: "Identifier", name: "promise" },
-            },
-          })
-          .forEach((callExpressionPath) => {
-            callExpressionPath.parentPath.value.object = (
-              callExpressionPath.value.callee as any
-            ).object;
-          });
-      });
-
-    replaceClientCreation(j, source, {
-      v2DefaultImportName,
-      v2ClientName,
-      v3ClientName,
-    });
+    removePromiseCalls(j, source, { v2DefaultImportName, v2ClientName });
+    replaceClientCreation(j, source, { v2DefaultImportName, v2ClientName, v3ClientName });
   }
 
   removeDefaultImportIfNotUsed(j, source, v2DefaultImportName);
