+++ src/transforms/v2-to-v3/utils/index.ts
@@ -3,5 +3,6 @@ export * from "./getClientMetadata";
 export * from "./getV2ClientNames";
 export * from "./getV2DefaultImportName";
 export * from "./removeDefaultImportIfNotUsed";
+export * from "./removePromiseCalls";
 export * from "./replaceClientCreation";
 export * from "./types";
